4.19.1 release

# Conflicts:
#	Engine/Build/Commit.gitdeps.xml
#	Engine/Config/ConsoleVariables.ini
#	Engine/Source/Runtime/D3D12RHI/Private/D3D12CommandContext.h
#	Engine/Source/Runtime/OpenGLDrv/Private/OpenGLCommands.cpp
#	Engine/Source/Runtime/RHI/Public/DynamicRHI.h
#	Engine/Source/Runtime/RHI/Public/RHICommandList.h
#	Engine/Source/Runtime/RHI/Public/RHICommandListCommandExecutes.inl
#	Engine/Source/Runtime/Windows/D3D11RHI/Private/D3D11Device.cpp

+++ Engine/Source/Developer/AutomationDriver/Private/StepExecutor.cpp
@@ -1,4 +1,4 @@
-// Copyright 1998-2017 Epic Games, Inc. All Rights Reserved.
+// Copyright 1998-2018 Epic Games, Inc. All Rights Reserved.
 
 #include "StepExecutor.h"
 #include "IStepExecutor.h"
@@ -133,7 +133,7 @@ private:
 		}
 
 		CurrentStepIndex = StepIndex;
-		float Milliseconds = Result.NextWait.GetTicks() / ETimespan::TicksPerMillisecond;
+		float Milliseconds = (float)(Result.NextWait.GetTicks()) / ETimespan::TicksPerMillisecond;
 		float Delay = FMath::Max(SMALL_NUMBER, (Milliseconds / 1000) * Configuration->ExecutionSpeedMultiplier);
 		
 		if (LastDelay < KINDA_SMALL_NUMBER)
