+++ Engine/Source/Developer/ShaderFormatVectorVM/Private/ShaderFormatVectorVM.cpp
@@ -1,4 +1,4 @@
-// Copyright 1998-2016 Epic Games, Inc. All Rights Reserved.
+// Copyright 1998-2018 Epic Games, Inc. All Rights Reserved.
 //
 
 #include "ShaderFormatVectorVM.h"
@@ -29,7 +29,7 @@ class FShaderFormatVectorVM : public IShaderFormat
 	}
 
 public:
-	virtual uint16 GetVersion(FName Format) const override
+	virtual uint32 GetVersion(FName Format) const override
 	{
 		CheckFormat(Format);
 		uint32 VVMVersion = 0;
